# Conflicts:
#	client/src/main/java/client/scenes/AddBoardController.java
#	client/src/main/java/client/scenes/AddListController.java
#	client/src/main/java/client/scenes/AddSubTaskController.java
#	client/src/main/java/client/scenes/EditBoardController.java
#	client/src/main/java/client/scenes/EditListController.java
#	client/src/main/java/client/scenes/EditTaskController.java
#	client/src/main/java/client/scenes/SubTaskShapeCtrl.java
#	client/src/main/java/client/scenes/TaskColorShape.java
#	client/src/main/java/client/scenes/tags/AddTagController.java
#	client/src/main/java/client/scenes/tags/AddTagToTaskController.java
#	client/src/main/java/client/utils/ServerUtils.java

+++ client/src/main/java/client/scenes/BoardController.java
@@ -49,7 +49,7 @@ public class BoardController {
     private ListShapeCtrl selectedList=null;
     private TaskShape selectedTask=null;
     private boolean editable = false;
-    private UserData userData;
+    private final UserData userData;
     private BoardService boardService;
     private TaskService taskService;
 
@@ -161,13 +161,23 @@ public class BoardController {
         showCtrl.showConnection();
     }
 
-    public void showEditBoard() { showCtrl.showEditBoard();}
+    public void showEditBoard() {
+        if(userData.isCurrentBoardLocked()){
+            userData.showError();
+            return;
+        }
+        showCtrl.showEditBoard();
+    }
 
     public void delete() {
+        if(userData.isCurrentBoardLocked()){
+            userData.showError();
+            return;
+        }
         IdResponseModel response = this.boardService.delete(
                 this.boardService.getCurrentBoard().getId());
 
-        if (response.getId() == -1) {
+        if (response.getId() < 0) {
             showCtrl.showError(response.getErrorMessage());
         }
         else{
